Allow removal of extension
This additional option allows the removal of the extension.
This is especially useful for TypeScript because the extension needs to be
removed.

This can be made more intelligent by allowing the user to specify removal
of the extension per language. For example, the configuration box would allow
a comma separated list of course matcher such as `source.ts, source.tsx`.

This would need to expose the current file source type. I will need to look into the
Atom APIs more to get this completed.